Create models file
